requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
openai==1.46.0
pypdf==4.2.0
docx2txt==0.8
python-docx==1.1.2
python-pptx==0.6.23
pillow==11.0.0
beautifulsoup4==4.12.3
lxml==5.3.0
python-multipart==0.0.9
requests==2.32.3
httpx==0.27.2
python-dotenv==1.0.1
gunicorn==23.0.0
asgiref==3.8.1
boto3==1.35.19
SQLAlchemy==2.0.35
psycopg[binary,pool]==3.2.10
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
